requirements.txt
selenium==4.16.0
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
python-dotenv==1.0.0
webdriver-manager==4.0.1
